// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface ExperienceDocumentDataBadgesItem {
  /**
   * Name field in *Experience → Badges*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.badges[].name
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  name: prismic.KeyTextField;

  /**
   * Color field in *Experience → Badges*
   *
   * - **Field Type**: Select
   * - **Placeholder**: *None*
   * - **Default Value**: whiteAlpha
   * - **API ID Path**: experience.badges[].color
   * - **Documentation**: https://prismic.io/docs/field#select
   */
  color: prismic.SelectField<
    | "whiteAlpha"
    | "blackAlpha"
    | "gray"
    | "red"
    | "orange"
    | "yellow"
    | "green"
    | "teal"
    | "blue"
    | "cyan"
    | "purple"
    | "pink",
    "filled"
  >;
}

/**
 * Item in *Experience → Tags*
 */
export interface ExperienceDocumentDataTagsItem {
  /**
   * Label field in *Experience → Tags*
   *
   * - **Field Type**: Select
   * - **Placeholder**: *None*
   * - **Default Value**: Frontend
   * - **API ID Path**: experience.tags[].label
   * - **Documentation**: https://prismic.io/docs/field#select
   */
  label: prismic.SelectField<"Frontend" | "Backend", "filled">;
}

/**
 * Content for Experience documents
 */
interface ExperienceDocumentData {
  /**
   * Logo field in *Experience*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.logo
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  logo: prismic.ImageField<never>;

  /**
   * Company field in *Experience*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.company
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  company: prismic.KeyTextField;

  /**
   * Position field in *Experience*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.position
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  position: prismic.KeyTextField;

  /**
   * Started In field in *Experience*
   *
   * - **Field Type**: Date
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.started_in
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#date
   */
  started_in: prismic.DateField;

  /**
   * Finished In field in *Experience*
   *
   * - **Field Type**: Date
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.finished_in
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#date
   */
  finished_in: prismic.DateField;

  /**
   * Badges field in *Experience*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.badges[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  badges: prismic.GroupField<Simplify<ExperienceDocumentDataBadgesItem>>;

  /**
   * Responsibilities field in *Experience*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.responsibilities
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  responsibilities: prismic.RichTextField;

  /**
   * Tags field in *Experience*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: experience.tags[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  tags: prismic.GroupField<Simplify<ExperienceDocumentDataTagsItem>>;
}

/**
 * Experience document from Prismic
 *
 * - **API ID**: `experience`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type ExperienceDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<ExperienceDocumentData>,
    "experience",
    Lang
  >;